Help browsers avoid fetching unused "Droid Sans" font.
AFAICT fonts shouldn't be loaded unless they're actually
used, but despite not being used[1] listing the font
here causes the browser to fetch this font.

[1] This font is used by the bootstrap css but
those values are not used due to local specification
to use Open Sans instead.

Anyway, I imagine there are better explanations for all
this but near as I can tell this changes nothing about
the behavior of the website and saves folks from
pulling down a font they won't be using :).

It's likely in their cache anyway, but it's the
spirit of the thing :).

This change does mean Droid Sans won't be used as
an explicit fallback, but the circumstances
where that behavior would be relied on seem unlikely.